decompose bot
